[quant][pt2] Add Conv + BN fusion for prepare QAT
Summary: This commit adds the prepare_qat_pt2e API and the
fusion logic for Conv + BN. We use the subgraph rewriter to
match and replace the pattern with the existing logic in
`nniqat.ConvBn2d`. Note this is not the end-to-end flow yet.
In particular, the following things are still missing and
will be added in future PRs:

1) Weight fake quantize after scaled weight
2) Convert flow needs to swap the new subgraph with another
   one that merges the batchnorm stats back into conv

Test Plan:
python test/test_quantization.py TestQuantizePT2E.test_qat_conv_bn_fusion

ghstack-source-id: d2374acdd7b86deceff4932f7a2f5ad4b99bc308
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98568